Organize Textract modules under ocr/textract folder
