+++ .gitignore
@@ -2,7 +2,7 @@
 .env.*
 *.pem
 /models/
-.DS_Store/
+.DS_Store
 __pycache__/
 *.pt
 *.bin
