+++ ai_models/requirements.txt
@@ -4,8 +4,6 @@ transformers
 torch
 pydantic
 huggingface_hub
-transformers
-torch
 protobuf>=3.20
 sentencepiece
 python-dotenv
