+++ app/interfaces/ocr_service.py
