+++ app/models/ocr_response.py
@@ -4,3 +4,4 @@ from typing import Dict, Union
 
 class OCRResponse(BaseModel):
     fields: Dict[str, Union[str, Dict[str, str]]]
+
