+++ app/services/ocr/textract/extractor.py
@@ -1,0 +1,47 @@
+from typing import Dict, List
+
+
+class TextractExtractor:
+    def extract_fields(self, blocks: List[dict]) -> Dict[str, str]:
+        key_map = {}
+        value_map = {}
+        block_map = {}
+
+        for block in blocks:
+            block_id = block["Id"]
+            block_map[block_id] = block
+            if block["BlockType"] == "KEY_VALUE_SET":
+                if "KEY" in block.get("EntityTypes", []):
+                    key_map[block_id] = block
+                elif "VALUE" in block.get("EntityTypes", []):
+                    value_map[block_id] = block
+
+        field_dict = {}
+        for key_block_id, key_block in key_map.items():
+            key_text = self._get_text(key_block, block_map)
+            value_text = ""
+
+            for relationship in key_block.get("Relationships", []):
+                if relationship["Type"] == "VALUE":
+                    for value_id in relationship["Ids"]:
+                        value_block = value_map.get(value_id)
+                        if value_block:
+                            value_text = self._get_text(value_block, block_map)
+
+            if key_text:
+                field_dict[key_text] = value_text
+
+        return field_dict
+
+    def _get_text(self, block: dict, block_map: dict) -> str:
+        text = ""
+        for rel in block.get("Relationships", []):
+            if rel["Type"] == "CHILD":
+                for child_id in rel["Ids"]:
+                    word = block_map[child_id]
+                    if word["BlockType"] == "WORD":
+                        text += word["Text"] + " "
+                    elif word["BlockType"] == "SELECTION_ELEMENT" and word["SelectionStatus"] == "SELECTED":
+                        text += "[X] "
+        return text.strip()
+
